# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Autoconf Vehicles API
  version: 1.0.0
  description: API stateless para usuários, veículos e imagens.
servers:
  - url: http://localhost:8000/api
security:
  - bearerAuth: []
paths:
...
  /vehicles/{vehicle}:
    parameters:
      - {name: vehicle, in: path, required: true, schema: {type: integer}}
    get:
      summary: Exibe veículo
      responses: {'200': {description: Veículo}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
    put:
      summary: Substitui veículo
      requestBody: {$ref: '#/components/requestBodies/Vehicle'}
      responses: {'200': {description: Atualizado}, '403': {$ref: '#/components/responses/Forbidden'}, '422': {$ref: '#/components/responses/Invalid'}}
    patch:
      summary: Atualiza parcialmente o veículo
      requestBody: {$ref: '#/components/requestBodies/Vehicle'}
      responses: {'200': {description: Atualizado}, '403': {$ref: '#/components/responses/Forbidden'}, '422': {$ref: '#/components/responses/Invalid'}}
    delete:
      summary: Desativa veículo
      responses: {'204': {description: Desativado}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: Sanctum PAT}
  schemas:
...
    VehicleInput:
      type: object
      required: [placa, chassi, marca, modelo, versao, valor_venda, cor, km, cambio, combustivel]
      properties:
        placa: {type: string, pattern: '^[A-Za-z]{3}[0-9][A-Za-z][0-9]{2}$'}
        chassi: {type: string, minLength: 17, maxLength: 17}
        marca: {type: string}
        modelo: {type: string}
        versao: {type: string}
        valor_venda: {type: number, minimum: 0.01}
        cor: {type: string}
        km: {type: integer, minimum: 0}
        cambio: {type: string, enum: [manual, automatico]}
        combustivel: {type: string, enum: [gasolina, alcool, flex, diesel, hibrido, eletrico]}
...
    Problem:
      type: object
      required: [type, title, status, message]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        message: {type: string}
        errors: {type: object, additionalProperties: {type: array, items: {type: string}}}
  requestBodies:
...
    Vehicle:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/VehicleInput'}}}
  responses:
...
    Forbidden: {description: Operação não permitida, content: {application/json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    NotFound: {description: Recurso não encontrado, content: {application/json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Invalid: {description: Dados inválidos, content: {application/json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...